mteb_layer_aggregation/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
mteb>=2.0.0
sentence-transformers>=2.2.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Optimization
cvxpy>=1.3.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
